+++ cli.go
@@ -76,7 +76,7 @@ COPYRIGHT:
 	app := cli.NewApp()
 
 	app.Name = "Oglang"
-	app.Version = "DEV"
+	app.Version = "v0.7.1"
 	app.Compiled = time.Now()
 
 	app.Usage = "Golang on steroids"